pyproject.toml
[tool.poetry]
name = "petco-demo"
version = "0.1.0"
description = ""
authors = ["MindInSky <kevin.a.muniz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.15.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
